Changed the getOrderFields method
BEFORE -> was just an simple array with fields allowed to be ordered with.

This is OK for simple usages, but not for ordering on complex cases like agregations (COUNT, etc..)

AFTER -> getOrderField is an associative array, with key as "property_name" passed to the api and value as the real SQL expression.